src/main/java/Student.java
import java.util.ArrayList;

public class Student {

    private String name;
    private int id;
    private ArrayList<Integer> grades;

    public Student(String name, int id) {
        this.name = name;
        this.id = id;
        grades = new ArrayList<Integer>();
    }

    public String getName() {
        return this.name;
    }

    public int getId() {
        return this.id;
    }

    public ArrayList<Integer> getGrades() {
        return this.grades;
    }

    public void addGrade(int grade) {
        grades.add(grade);
    }

    public double getGradeAverage() {
        double sum = 0;
        for(double n : this.grades) {
            sum+=n;
        }
        return sum / this.grades.size();
    }
}
